Gcc 14 clang tidy 21 (#271)
* update gcc to 14.2.0

* update clang-tidy to 21.1.8

* added clang-tidy bugprone-* diagnostics

* added metrics folder to clang-tidy check

* commit for test pipeline

* Separate CI image for tests


* fix queue


* fixed compilation error

* force using of gcc14


* fake commit


* test


* test


* revert


* fixed compilation error

* removed bazel cache (temporary)

* reverted commit for remove bazel caches

* remove uploaded logs


* fixed copy-paste bug

* review fixes

* add registry cleanup


---------

+++ pp/performance_tests/load_gorilla_from_wal_and_make_remote_write_from_it_test.cpp
@@ -29,5 +29,5 @@ std::chrono::nanoseconds load_gorilla_from_wal_and_make_remote_write_from_it::pr
 }
 
 void load_gorilla_from_wal_and_make_remote_write_from_it::write_metrics(Metrics& metrics) const {
-  metrics << (Metric() << "wal_loading_from_file_and_make_remote_write_duration_microseconds" << (period_.count() / 1000.0));
+  metrics << (Metric() << "wal_loading_from_file_and_make_remote_write_duration_microseconds" << static_cast<double>(period_.count()) / 1000.0);
 }
